lib/logstorage: replace the first row in a block with an empty row with the original timestamp if it is deleted during background merge
VictoriaLogs expects tha data blocks inside parts are sorted by (streamID, timestamp),
where timestamp is the minimum timestamp in the block. Rows are sorted by timestamps inside the block,
so the first row in the block has the minimum timestamp.

If the first row is deleted in the data block, then this may lead to the violation of the expectations during
merge of the blocks, so the merge could produce output blocks, which aren't sorted by the minimum timestamp
per block. This could lead to data loss for the subsequent queries and merges of the produced blocks,
since these code paths rely on the assumption that blocks inside parts are sorted by (streamID, timestamp).

The potential data loss was prevented by the check inside blockStreamMerger.checkNextBlock(),
which tirggers the fatal error at block_stream_merger.go:237.

The solution is to replace the first row in a block with an empty row if this row is deleted, while preserving its' original timestamp.
This prevents from the violation of the expectations that blocks inside parts are sorted by (streamID, timestamp).
An empty row is invisible for LogsQL filters thanks to VictoriaLogs data model, which treats empty fields as non-existing fields.
See https://docs.victoriametrics.com/victorialogs/keyconcepts/#data-model

Update TestStorageProcessDeleteTask(), so it uses count(host) instead of count() for counting the number of matching rows,
since count() counts all the rows, even if all the fields are empty in these rows. This is an optimization, which cannot be removed.
The count(host) counts rows with non-empty host field, so it doesn't take into account empty rows without fields.
See https://docs.victoriametrics.com/victorialogs/logsql/#count-stats for more details.

Updates https://github.com/VictoriaMetrics/VictoriaLogs/issues/825
Updates https://github.com/VictoriaMetrics/VictoriaLogs/pull/827